requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.44
pypdf==6.6.2
langchain>=1.2.0
langchain-openai>=1.0.0
langchain-community>=0.4.0
langchain-pinecone>=0.2.0
pinecone>=6.0.0
python-dotenv==1.0.1
reportlab==4.4.9
python-multipart==0.0.20
python-docx==1.1.2
cryptography>=3.1
langchain-google-genai>=1.0.0
